sessions_mount_dir/0pfl408b6hw9t4le87vmki41a/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
prometheus-client==0.19.0
